dynamic routes: relocates pages from pages/myrw
