+++ components/app/myrw/datasets/DatasetsTab.js
@@ -14,6 +14,7 @@ function DatasetsTab(props) {
     tab, subtab, id, user,
   } = props;
 
+
   return (
     <div className="c-datasets-tab">
       {!id && user.token
