+++ components/app/myrw/widgets/tabs/list/constants.js
@@ -2,13 +2,13 @@ export const WIDGET_LIST_SUBTABS = [
   {
     label: 'My visualizations',
     value: 'my_wigets',
-    route: 'myrw',
+    route: '/myrw/widgets/my_widgets',
     params: { tab: 'widgets', subtab: 'my_widgets' },
   },
   {
     label: 'Favorites',
     value: 'favourites',
-    route: 'myrw',
+    route: '/myrw/widgets/favorites',
     params: { tab: 'widgets', subtab: 'favorites' },
   },
 ];
