+++ layout/myrw/detail/index.js
@@ -9,7 +9,6 @@ import LayoutMyRWDetail from './component';
 export default connect(
   (state) => ({
     user: state.user,
-    query: state.routes.query,
     locale: state.common.locale,
     myrwdetail: state.myrwdetail,
     alerts: areaAlerts(state),
